+++ navigation2/package.xml
@@ -16,7 +16,7 @@
   <depend>nav2_astar_planner</depend>
   <depend>nav2_bt_navigator</depend>
   <depend>nav2_costmap_2d</depend>
-  <depend>nav2_costmap_world_model</depend>
+  <depend>nav2_world_model</depend>
   <depend>nav2_dijkstra_planner</depend>
   <depend>nav2_gazebo_localizer</depend>
   <depend>nav2_map_server</depend>
